imp: add Confloader
Make conf loading reusable
