+++ oca_repo_maintainer/tools/manager.py
@@ -4,21 +4,18 @@
 # @author: Simone Orsi
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
 
-import hashlib
 import logging
 import shutil
 import subprocess
 import sys
 import tempfile
-from pathlib import Path
 from subprocess import CalledProcessError
 
 import copier
 import github3
-import yaml
 from github3.exceptions import NotFoundError
 
-from .utils import SmartDict
+from .utils import ConfLoader
 
 handler = logging.StreamHandler(sys.stdout)
 handler.setLevel(logging.INFO)
@@ -52,72 +49,21 @@ class RepoManager:
     """Setup and update repositories and teams."""
 
     def __init__(self, conf_dir, org, token, force=False):
-        self.conf_dir = Path(conf_dir)
+        self.conf_dir = conf_dir
+        self.conf_loader = ConfLoader(self.conf_dir)
         self.token = token
         self.org = org
         self.force = force
-        self.checksum = {}
-        if not force:
-            self.checksum = self._load_conf("checksum", checksum=False)
-        self.conf_global = self._load_conf("global")
-        self.conf_psc = self._load_conf("psc")
-        self.conf_repo = self._load_conf("repo")
+        self.conf_global = self.conf_loader.load_conf("global", checksum=False)
+        self.conf_psc = self.conf_loader.load_conf("psc", checksum=not force)
+        self.conf_repo = self.conf_loader.load_conf("repo", checksum=not force)
         self.new_repo_template = self.conf_global.get("template")
-
-    def _load_conf(self, name, checksum=True):
-        conf = {}
-        path = self.conf_dir / name
-        filepath = path.with_suffix(".yml")
-        if filepath.exists():
-            # direct yml files
-            conf.update(self._load_conf_from_file(filepath, checksum=checksum))
-        else:
-            # folders containing ymls
-            for filepath in path.rglob("*.yml"):
-                conf.update(self._load_conf_from_file(filepath))
-        return SmartDict(conf)
-
-    def _load_conf_from_file(self, filepath, checksum=True):
-        conf = {}
-        with filepath.open() as fd:
-            content = fd.read()
-            if not content:
-                return conf
-            if checksum and self._file_changed(filepath, content):
-                conf.update(yaml.safe_load(content))
-                self._store_checksum(filepath, content)
-            elif not checksum:
-                conf.update(yaml.safe_load(content))
-            else:
-                _logger.info(
-                    "%s not changed: skipping", self._filepath_for_checksum(filepath)
-                )
-        return conf
-
-    def _file_changed(self, filepath, content):
-        return self._make_md5(content) != self.checksum.get(
-            self._filepath_for_checksum(filepath)
-        )
-
-    def _make_md5(self, content):
-        return hashlib.md5(content.encode()).hexdigest()
-
-    def _save_checksum(self):
-        if self.checksum:
-            with open(self.conf_dir / "checksum.yml", "w") as f:
-                yaml.dump(dict(self.checksum), f)
-
-    def _store_checksum(self, filepath, content):
-        self.checksum[self._filepath_for_checksum(filepath)] = self._make_md5(content)
-
-    def _filepath_for_checksum(self, filepath):
-        return filepath.relative_to(self.conf_dir).as_posix()
 
     def run(self):
         self._setup_gh()
         self._process_psc()
         self._process_repositories()
-        self._store_checksum()
+        self._save_checksum()
 
     def _setup_gh(self):
         self.gh = github3.login(token=self.token)
@@ -145,7 +91,7 @@ class RepoManager:
                 cwd=clone_dir,
             )
 
-    def _process_psc(self, foo=0):
+    def _process_psc(self):
         if not self.conf_psc:
             _logger.info("No team to process")
             return
@@ -288,3 +234,6 @@ class RepoManager:
 
     def _run_cmd(self, cmd, cwd, **kw):
         check_call(cmd, cwd, **kw)
+
+    def _save_checksum(self):
+        self.conf_loader.save_checksum()
