--- tests/conf/checksum.yml
@@ -1,5 +1,0 @@
-global.yml: e216cad006500594046936677dc928bd
-psc/psc1.yml: 808bcdc95335aeb5c960cdcb93bf7712
-psc/psc2.yml: 903910e15be9433d8472183ea91b99f2
-repo/repo1.yml: 95d21fa912ef332efcb544bdbe256414
-repo/repo2.yml: 93159b31158e26d225086d3b0fac7289
