+++ tests/conf2/checksum.yml
@@ -1,0 +1,4 @@
+psc/psc1.yml: 808bcdc95335aeb5c960cdcb93bf7712
+psc/psc2.yml: 903910e15be9433d8472183ea91b99f2
+repo/repo1.yml: 245cbc6100109f329daaed65c7522c68
+repo/repo2.yml: 5e92d2f7731d70481c8bedc6ec7f77ba
