+++ tests/conf2/psc/psc1.yml
@@ -1,0 +1,6 @@
+test-team-1:
+  name: Test team 1
+  members:
+    - simahawk
+  representatives:
+    - simahawk
